Refactored Static Component Getter to Costom Components to reuse them outside forms view, cleanup
