+++ taz.neo/ViewController/Forms/TrialSubscriptionController.swift
@@ -52,25 +52,25 @@ class TrialSubscriptionController : FormsController {
   
   override func getContentViews() -> [UIView] {
     return  [
-      FormularView.header(),
-      FormularView.label(title: Localized("trial_subscription_title")),
+      TazHeader(),
+      UILabel(title: Localized("trial_subscription_title")),
       mailInput,
       passInput,
       pass2Input,
       firstnameInput,
       lastnameInput,
-      FormularView.label(title:
+      UILabel(title:
         Localized("fragment_login_request_test_subscription_existing_account")),
       //      FormularView.labelLikeButton(title: Localized("login_forgot_password"),
       //                                   target: self,
       //                                   action: #selector(handlePwForgot)),
       contentView.agbAcceptTV,
-      FormularView.button(title: Localized("register_button"),
-                          target: self,
-                          action: #selector(handleSend)),
-      FormularView.outlineButton(title: Localized("cancel_button"),
-                                 target: self,
-                                 action: #selector(handleCancel)),
+      UIButton(title: Localized("register_button"),
+               target: self,
+               action: #selector(handleSend)),
+      UIButton(type: .outline, title: Localized("cancel_button"),
+               target: self,
+               action: #selector(handleCancel)),
     ]
   }
